requirements.txt
asgiref==3.2.7
Django==3.0.4
djangorestframework==3.11.0
geojson==2.5.0
numpy==1.18.2
pytz==2019.3
Shapely==1.7.0
sqlparse==0.3.1
